chore: bail out of lazychild earlier
