+++ example/package.json
@@ -9,18 +9,18 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "expo": "~47.0.12",
-    "expo-status-bar": "~1.4.2",
+    "expo": "^51.0.37",
+    "expo-status-bar": "~1.12.1",
     "lodash": "^4.17.21",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-native": "0.70.5",
-    "react-native-reanimated": "~2.12.0",
-    "react-native-web": "~0.18.9"
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-native": "0.74.5",
+    "react-native-reanimated": "~3.10.1",
+    "react-native-web": "~0.19.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@expo/webpack-config": "^0.17.2",
+    "@babel/core": "^7.24.0",
+    "@expo/webpack-config": "~19.0.1",
     "@types/lodash": "^4.14.194",
     "babel-loader": "^8.1.0",
     "babel-plugin-module-resolver": "^4.1.0"
