+++ package.json
@@ -71,7 +71,7 @@
     "react": "18.1.0",
     "react-native": "0.70.5",
     "react-native-builder-bob": "^0.20.3",
-    "react-native-reanimated": "^2.14.4",
+    "react-native-reanimated": "^3.15.5",
     "react-test-renderer": "^18.2.0",
     "release-it": "^15.0.0",
     "typedoc": "^0.26.7",
@@ -81,9 +81,9 @@
     "@types/react": "17.0.21"
   },
   "peerDependencies": {
-    "react": "*",
-    "react-native": "*",
-    "react-native-reanimated": "^2.14.4"
+    "react": ">=18",
+    "react-native": ">=0.70",
+    "react-native-reanimated": ">=3"
   },
   "engines": {
     "node": ">= 16.0.0"
